refactor: moving the parser into the parse folder
